# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AEGIS Security System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AttackTypeStat {
  type: string;
  count: number;
  severity: string;
}

export interface TrendPoint {
  hour: string;
  count: number;
}

export interface DashboardSummary {
  totalEvents: number;
  criticalEvents: number;
  openIncidents: number;
  activeAlerts: number;
  blockedIPs: number;
  systemsOnline: number;
  systemsTotal: number;
  attacksByType: AttackTypeStat[];
  eventsTrend: TrendPoint[];
}

export type SecurityEventSeverity = typeof SecurityEventSeverity[keyof typeof SecurityEventSeverity];


export const SecurityEventSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type SecurityEventStatus = typeof SecurityEventStatus[keyof typeof SecurityEventStatus];


export const SecurityEventStatus = {
  detected: 'detected',
  blocked: 'blocked',
  investigating: 'investigating',
  resolved: 'resolved',
} as const;

export type SecurityEventLayer = typeof SecurityEventLayer[keyof typeof SecurityEventLayer];


export const SecurityEventLayer = {
  perimeter: 'perimeter',
  brain: 'brain',
  output: 'output',
} as const;

export interface SecurityEvent {
  id: number;
  type: string;
  subtype: string;
  severity: SecurityEventSeverity;
  sourceIp: string;
  targetHost: string;
  /** @nullable */
  toolUsed?: string | null;
  description: string;
  status: SecurityEventStatus;
  layer: SecurityEventLayer;
  createdAt: string;
}

export type SecurityEventInputSeverity = typeof SecurityEventInputSeverity[keyof typeof SecurityEventInputSeverity];


export const SecurityEventInputSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface SecurityEventInput {
  type: string;
  subtype: string;
  severity: SecurityEventInputSeverity;
  sourceIp: string;
  targetHost: string;
  toolUsed?: string;
  description: string;
  layer: string;
}

export type IncidentSeverity = typeof IncidentSeverity[keyof typeof IncidentSeverity];


export const IncidentSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type IncidentStatus = typeof IncidentStatus[keyof typeof IncidentStatus];


export const IncidentStatus = {
  open: 'open',
  investigating: 'investigating',
  contained: 'contained',
  resolved: 'resolved',
} as const;

export interface Incident {
  id: number;
  title: string;
  severity: IncidentSeverity;
  status: IncidentStatus;
  description: string;
  /** @nullable */
  responder?: string | null;
  /** @nullable */
  notes?: string | null;
  eventCount: number;
  createdAt: string;
  updatedAt: string;
}

export type IncidentInputSeverity = typeof IncidentInputSeverity[keyof typeof IncidentInputSeverity];


export const IncidentInputSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export interface IncidentInput {
  title: string;
  severity: IncidentInputSeverity;
  description: string;
  responder?: string;
}

export type IncidentUpdateStatus = typeof IncidentUpdateStatus[keyof typeof IncidentUpdateStatus];


export const IncidentUpdateStatus = {
  open: 'open',
  investigating: 'investigating',
  contained: 'contained',
  resolved: 'resolved',
} as const;

export interface IncidentUpdate {
  status?: IncidentUpdateStatus;
  responder?: string;
  notes?: string;
}

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type AlertChannel = typeof AlertChannel[keyof typeof AlertChannel];


export const AlertChannel = {
  telegram: 'telegram',
  dashboard: 'dashboard',
  email: 'email',
} as const;

export interface Alert {
  id: number;
  message: string;
  severity: AlertSeverity;
  channel: AlertChannel;
  acknowledged: boolean;
  /** @nullable */
  eventId?: number | null;
  createdAt: string;
}

export type SystemStatusLayer = typeof SystemStatusLayer[keyof typeof SystemStatusLayer];


export const SystemStatusLayer = {
  attacker: 'attacker',
  perimeter: 'perimeter',
  brain: 'brain',
  output: 'output',
} as const;

export type SystemStatusStatus = typeof SystemStatusStatus[keyof typeof SystemStatusStatus];


export const SystemStatusStatus = {
  online: 'online',
  offline: 'offline',
  warning: 'warning',
  unknown: 'unknown',
} as const;

export interface SystemStatus {
  id: number;
  component: string;
  layer: SystemStatusLayer;
  status: SystemStatusStatus;
  description: string;
  /** @nullable */
  metrics?: string | null;
  lastCheck: string;
}

export type ReportType = typeof ReportType[keyof typeof ReportType];


export const ReportType = {
  daily: 'daily',
  weekly: 'weekly',
  incident: 'incident',
  custom: 'custom',
} as const;

export type ReportFormat = typeof ReportFormat[keyof typeof ReportFormat];


export const ReportFormat = {
  html: 'html',
  pdf: 'pdf',
} as const;

export interface Report {
  id: number;
  title: string;
  type: ReportType;
  format: ReportFormat;
  summary: string;
  eventsCount: number;
  incidentsCount: number;
  generatedAt: string;
}

export type ReportInputType = typeof ReportInputType[keyof typeof ReportInputType];


export const ReportInputType = {
  daily: 'daily',
  weekly: 'weekly',
  incident: 'incident',
  custom: 'custom',
} as const;

export type ReportInputFormat = typeof ReportInputFormat[keyof typeof ReportInputFormat];


export const ReportInputFormat = {
  html: 'html',
  pdf: 'pdf',
} as const;

export interface ReportInput {
  title: string;
  type: ReportInputType;
  format: ReportInputFormat;
}

export type ListEventsParams = {
limit?: number;
offset?: number;
severity?: ListEventsSeverity;
type?: string;
};

export type ListEventsSeverity = typeof ListEventsSeverity[keyof typeof ListEventsSeverity];


export const ListEventsSeverity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
} as const;

export type ListAlertsParams = {
acknowledged?: boolean;
};