# lib/api-zod/src/generated/types/attackTypeStat.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AEGIS Security System API
 * OpenAPI spec version: 0.1.0
 */

export interface AttackTypeStat {
  type: string;
  count: number;
  severity: string;
}