# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AEGIS Security System API
 * OpenAPI spec version: 0.1.0
 */
import type { AttackTypeStat } from './attackTypeStat';
import type { TrendPoint } from './trendPoint';

export interface DashboardSummary {
  totalEvents: number;
  criticalEvents: number;
  openIncidents: number;
  activeAlerts: number;
  blockedIPs: number;
  systemsOnline: number;
  systemsTotal: number;
  attacksByType: AttackTypeStat[];
  eventsTrend: TrendPoint[];
}