# lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * AEGIS Security System API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  hour: string;
  count: number;
}